+++ src/Command/Router/ListRoutes.php
@@ -3,17 +3,9 @@ declare(strict_types=1);
 
 namespace Heptacom\HeptaConnect\Bridge\ShopwarePlatform\Command\Router;
 
-use Heptacom\HeptaConnect\Core\Portal\PortalStackServiceContainerFactory;
-use Heptacom\HeptaConnect\Portal\Base\Emission\Contract\EmitterContract;
-use Heptacom\HeptaConnect\Portal\Base\Emission\EmitterCollection;
-use Heptacom\HeptaConnect\Portal\Base\Exploration\Contract\ExplorerContract;
-use Heptacom\HeptaConnect\Portal\Base\Exploration\ExplorerCollection;
-use Heptacom\HeptaConnect\Portal\Base\Reception\Contract\ReceiverContract;
-use Heptacom\HeptaConnect\Portal\Base\Reception\ReceiverCollection;
-use Heptacom\HeptaConnect\Storage\Base\Contract\Repository\PortalNodeRepositoryContract;
-use Heptacom\HeptaConnect\Storage\Base\Contract\Repository\RouteRepositoryContract;
+use Heptacom\HeptaConnect\Storage\Base\Contract\RouteOverviewActionInterface;
+use Heptacom\HeptaConnect\Storage\Base\Contract\RouteOverviewCriteria;
 use Heptacom\HeptaConnect\Storage\Base\Contract\StorageKeyGeneratorContract;
-use Heptacom\HeptaConnect\Storage\ShopwareDal\StorageKey\PortalNodeStorageKey;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
@@ -23,88 +15,36 @@ class ListRoutes extends Command
 {
     protected static $defaultName = 'heptaconnect:router:list-routes';
 
-    private RouteRepositoryContract $routeRepository;
-
-    private PortalNodeRepositoryContract $portalNodeRepository;
-
     private StorageKeyGeneratorContract $storageKeyGenerator;
 
-    private PortalStackServiceContainerFactory $portalStackServiceContainerFactory;
+    private RouteOverviewActionInterface $routeOverviewAction;
 
     public function __construct(
-        RouteRepositoryContract $routeRepository,
-        PortalNodeRepositoryContract $portalNodeRepository,
         StorageKeyGeneratorContract $storageKeyGenerator,
-        PortalStackServiceContainerFactory $portalStackServiceContainerFactory
+        RouteOverviewActionInterface $routeOverviewAction
     ) {
         parent::__construct();
-        $this->routeRepository = $routeRepository;
-        $this->portalNodeRepository = $portalNodeRepository;
         $this->storageKeyGenerator = $storageKeyGenerator;
-        $this->portalStackServiceContainerFactory = $portalStackServiceContainerFactory;
+        $this->routeOverviewAction = $routeOverviewAction;
     }
 
     protected function execute(InputInterface $input, OutputInterface $output)
     {
         $io = new SymfonyStyle($input, $output);
-        $types = [];
-
-        foreach ($this->portalNodeRepository->listAll() as $portalNodeKey) {
-            $container = $this->portalStackServiceContainerFactory->create($portalNodeKey);
-
-            /** @var ExplorerCollection $explorers */
-            $explorers = $container->get(ExplorerCollection::class);
-            /** @var ExplorerCollection $explorerDecorators */
-            $explorerDecorators = $container->get(ExplorerCollection::class.'.decorator');
-            $explorers->push($explorerDecorators);
-
-            /** @var EmitterCollection $emitters */
-            $emitters = $container->get(EmitterCollection::class);
-            /** @var EmitterCollection $emitterDecorators */
-            $emitterDecorators = $container->get(EmitterCollection::class.'.decorator');
-            $emitters->push($emitterDecorators);
-
-            /** @var ReceiverCollection $receivers */
-            $receivers = $container->get(ReceiverCollection::class);
-            /** @var ReceiverCollection $receiverDecorators */
-            $receiverDecorators = $container->get(ReceiverCollection::class.'.decorator');
-            $receivers->push($receiverDecorators);
-
-            /** @var ExplorerContract $explorer */
-            foreach ($explorers as $explorer) {
-                $types[$explorer->supports()] = true;
-            }
-
-            /** @var EmitterContract $emitter */
-            foreach ($emitters as $emitter) {
-                $types[$emitter->supports()] = true;
-            }
-
-            /** @var ReceiverContract $receiver */
-            foreach ($receivers as $receiver) {
-                $types[$receiver->supports()] = true;
-            }
-        }
-
-        $types = \array_keys($types);
         $targets = [];
 
-        foreach ($this->portalNodeRepository->listAll() as $portalNodeKey) {
-            if (!$portalNodeKey instanceof PortalNodeStorageKey) {
-                continue;
-            }
+        $criteria = new RouteOverviewCriteria();
+        $criteria->setSort([
+            RouteOverviewCriteria::FIELD_ENTITY_TYPE => RouteOverviewCriteria::SORT_ASC,
+            RouteOverviewCriteria::FIELD_CREATED => RouteOverviewCriteria::SORT_DESC,
+        ]);
 
-            foreach ($types as $type) {
-                foreach ($this->routeRepository->listBySourceAndEntityType($portalNodeKey, $type) as $routeKey) {
-                    $route = $this->routeRepository->read($routeKey);
-
-                    $targets[] = [
-                        'type' => $type,
-                        'source' => $this->storageKeyGenerator->serialize($portalNodeKey),
-                        'target' => $this->storageKeyGenerator->serialize($route->getTargetKey()),
-                    ];
-                }
-            }
+        foreach ($this->routeOverviewAction->overview($criteria) as $route) {
+            $targets[] = [
+                'type' => $route->getEntityType(),
+                'source' => $this->storageKeyGenerator->serialize($route->getSource()),
+                'target' => $this->storageKeyGenerator->serialize($route->getTarget()),
+            ];
         }
 
         if (\count($targets) === 0) {